\name{Consist}
\docType{data}
\alias{Consist}
\title{irrNA example data, showing perfect consistency between raters}
\description{This data set shows perfect consistency and moderate agreement between raters.}
\usage{data(Consist)}
\format{A 2-dimensional data frame including column and row headers.}
\keyword{datasets}