\name{ConsistNA}
\docType{data}
\alias{ConsistNA}
\title{irrNA example data, showing perfect consistency between raters and NAs}
\description{This data set shows missing values (NAs) and perfect consistency and moderate agreement between raters.}
\usage{data(ConsistNA)}
\format{A 2-dimensional data frame including column and row headers and NAs.}
\keyword{datasets}