\name{Ebel51}
\docType{data}
\alias{Ebel51}
\title{Example data, given by Ebel (1951, Table 2)}
\description{This data set is used by Ebel (1951) to demonstate the computation of an intraclass correlation on incomplete data sets.}
\usage{data(Ebel51)}
\format{A 2-dimensional data frame including column and row headers and NAs.}
\source{Psychometrika}
\references{Ebel, R.L. (1951). Estimation of the reliability of ratings. Psychometrika, 16(4), 407--424.}
\keyword{datasets}