\name{EbelFILL}
\docType{data}
\alias{EbelFILL}
\title{Example data, based on Ebel (1951, Table 2)}
\description{This data set is the same as Ebel51, but with the missing data filled up with arbitrary values.}
\usage{data(EbelFILL)}
\format{A 2-dimensional data frame including column and row headers.}
\references{Ebel, R.L. (1951). Estimation of the reliability of ratings. Psychometrika, 16(4), 407--424.}
\keyword{datasets}