\name{Indep}
\docType{data}
\alias{Indep}
\title{irrNA example data, showing perfect independence among raters and objects}
\description{This data set shows perfect independance among raters and objects.}
\usage{data(Indep)}
\format{A 2-dimensional data frame including column and row headers.}
\keyword{datasets}