\name{IndepNA}
\docType{data}
\alias{IndepNA}
\title{irrNA example data, showing NAs and perfect independence among raters and objects}
\description{This data set shows missing values (NAs) and perfect independance among raters and objects.}
\usage{data(IndepNA)}
\format{A 2-dimensional data frame including column and row headers and NAs.}
\keyword{datasets}