\name{IndepW}
\docType{data}
\alias{IndepW}
\title{irrNA example data, showing perfect independence among raters and NAs}
\description{This data set shows missing values (NAs) and perfect independance among raters.}
\usage{data(IndepW)}
\format{A 2-dimensional data frame including column and row headers and NAs.}
\keyword{datasets}